requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.6
cffi==1.14.5
chardet==4.0.0
discord==1.0.1
discord.py==1.7.1
idna==3.1
multidict==5.1.0
pycodestyle==2.7.0
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.15.0
six==1.16.0
toml==0.10.2
typing-extensions==3.7.4.3
yarl==1.6.3
youtube-dl==2021.2.10
